+++ src/layouts/main/footer.tsx
@@ -21,7 +21,7 @@ const LINKS = [
   {
     headline: 'Enabled',
     children: [
-      { name: 'About', href: paths.about },
+      { name: 'Features', href:'#' },
       { name: 'FAQs', href: paths.faqs },
     ],
   },
@@ -97,7 +97,7 @@ export default function Footer() {
                 mx: { xs: 'auto', md: 'unset' },
               }}
             >
-              Disability accommodations in the workplace made easy with EnableD.
+              Reasonable adjustments in the workplace made easy with enableD.
             </Typography>
 
             <Stack
