+++ src/sections/home/home-faq.tsx
